+++ README.md
@@ -36,8 +36,6 @@ This repository contains algorithms and data structures implemented in C# for ed
 		* [Binary](./Algorithms/Search/BinarySearcher.cs)
 		* [Linear](./Algorithms/Search/LinearSearcher.cs)
 		* [FastSearch](./Algorithms/Search/FastSearcher.cs)
-		* [Substring](./Algorithms/Search/Substring)
-			* [Rabin Karp](./Algorithms/Search/Substring/RabinKarp.cs)
 	* [Sorts](./Algorithms/Sorters/)
 		* [Comparison](./Algorithms/Sorters/Comparison)
 			* [Binary Insertion](./Algorithms/Sorters/Comparison/BinaryInsertionSorter.cs)
@@ -65,6 +63,7 @@ This repository contains algorithms and data structures implemented in C# for ed
 	* [String](./Algorithms/Strings/)
 		* [Longest Consecutive Character](./Algorithms/Strings/GeneralStringAlgorithms.cs)
 		* [Naive String Search](./Algorithms/Strings/NaiveStringSearch.cs)
+		* [Rabin Karp](./Algorithms/Strings/RabinKarp.cs)
 		* [Palindrome Checker](./Algorithms/Strings/palindrome.cs)
 	* [Other](./Algorithms/Other/)
 		* [Fermat Prime Checker](./Algorithms/Other/FermatPrimeChecker.cs)
